assets/js/requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.68.1
h11==0.12.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.0.3
numpy==1.21.2
opencv-contrib-python==4.2.0.34
opencv-python
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.1.1
pyzmq==22.2.1
requests==2.26.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
starlette==0.14.2
terminado==0.12.1
testpath==0.5.0
torch==1.9.0
torchaudio==0.9.0
torchfile==0.1.0
torchvision==0.10.0
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.1
zipp==3.5.0
